pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autofic-core"
version = "0.1.0"
description = "A solution for remediating vulnerable source code using LLMs."
authors = [
  {name = "autofic", email = "autofic.whs@gmail.com"},
  {name = "Minchae Kim", email = "minxxcozy@gmail.com"},
  {name = "Eunsol Kim"},
  {name = "Seonju Park"},
  {name = "Jeongmin Oh"},
  {name = "Inyeong Jang", email = "726jang@gmail.com"},
  {name = "Hongseo Jang", email = "janghongseo@naver.com"},
  {name = "Yunji Jeong"},
  {name = "Yunjeong Choe"},
  {name = "Suhyun Park", email = "lovehyun@hotmail.com"},
  {name = "Changhyun Lee", email = "dlckdgus200011@gmail.com"},
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.3,<9.0.0",
    "requests>=2.28.1,<3.0.0",
    "python-dotenv>=0.21.0,<1.0.0",
    "rich>=13.3.1,<14.0.0",
    "openai>=1.0.0,<2.0.0",
    "pyfiglet>=1.0.3,<2.0.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
]
keywords = ["llm", "security", "autofix", "static analysis", "code remediation"]

[project.scripts]
autofic-core = "autofic_core.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]